Usability features (#20)
* Rename TClaim to TrustClaim

The new name is both, consistent with TrustTier, and clearer.

* Add TrustClaim constants

Add constants definitions for the defined claim values across trust
vector claim categories.

* Usability features

Implement a number of changes and additions to aid in usability of this
library. No material changes are being made toe the forts of EAR or
AR4SI as part of this.

- Rename To/FromJSON to Marshal/UnmarshaJSON to be consistent with
  common golang conventions. Rename ToJSONPretty to MarshalJSONIndent
  for the same reason.
- Add AsMap() methods to TrustVector and AttestationResult that converts
  these structs to map[string]interface{}.
- Add UpdateStatusFromTrustVector() to AttestationResult, that brings
  Status into alignment with TrustVector values (unless it was
  explicitly set to a lower trust value).
- Add ToTrustTier and ToTrustClaim to convert arbitrary interfaces to
  corresponding structs.
- Add TrustClaim.GetTier() that returns the tier corresponding to the
  claim value.
- Add NewAttestationResult that returns a fully initialized attestation
  result.
- Switch to using jwt package for signing and verifying.

* Lower coverage requirements to 80%

Previous commit introduces an number of type switches with multitudes of
identical cases for all the different integer types. This technically lowers
statement coverage of the tests without materially affecting logical
coverage of the code.

Signed-off-by: setrofim <setrofim@gmail.com>